Change Eta property type from long to double for improved precision in download metrics

+++ Domain/DTOs/DownloadItem.cs
@@ -7,7 +7,7 @@ public record DownloadItem : SearchResult
     public required double Progress { get; init; }
     public required double DownSpeed { get; init; }
     public required double UpSpeed { get; init; }
-    public required long Eta { get; init; }
+    public required double Eta { get; init; }
 }
 
 public enum DownloadStatus
